+++ OneMore/App.config
@@ -31,7 +31,7 @@
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="System.Threading.Tasks.Extensions" publicKeyToken="cc7b13ffcd2ddd51" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-9.9.9.9" newVersion="4.2.4.0" />
+        <bindingRedirect oldVersion="0.0.0.0-9.9.9.9" newVersion="4.2.0.1" />
       </dependentAssembly>
     </assemblyBinding>
   </runtime>
