+++ OneMoreCalendar/App.config
@@ -32,7 +32,7 @@
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
       <dependentAssembly>
         <assemblyIdentity name="System.Threading.Tasks.Extensions" publicKeyToken="cc7b13ffcd2ddd51" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-4.2.4.0" newVersion="4.2.4.0" />
+        <bindingRedirect oldVersion="0.0.0.0-4.2.4.0" newVersion="4.2.0.1" />
       </dependentAssembly>
     </assemblyBinding>
   </runtime>
